Solves validation bug by adding calls to init<..>Label functions
Add calls to init<..>Label functions to preUpdateConversions()
and to setId<..>Label and set<..>Label.

Without these calls isDataOK() could not validate the contents
of the labels if they were passed directly instead of via
set<..>Label(String, DbBeanLanguage).

The calls in preUpdateConversions prevents a validation bug if
updateDB() is called without modifying one of the label.