Recode support for transactions in bean updating
Makes it easier to write procedural code that updates beans in a
DBTransaction.

Also solves a bug when checking the validity of a bean's ID when
the bean is new and has only been saved to the database in the
context of the transaction and the records are not yet committed.