Add variables formName, horizontal & readonly to <Bean>HTMLViewBase
formName allows the name of the form to be referenced from one place
instead of 3 (simplifies creation of derived forms with a different
name).

horizontal: to indicate if the form should have an horizontal layout
(it's no longer needed to override getFormStart() for this).

readonly: the form will be rendered with all fields marked readonly.